Refactor kafka into separate resources, add recalculate embargo consumer
