--- src/types/kafkaMessages.ts
@@ -1,11 +1,0 @@
-export type PublicReleaseMessage = {
-  id: string;
-  publishedAt: Date;
-  label: string;
-  programs: Program[];
-};
-
-export type Program = {
-  id: string;
-  donorsUpdated: string[];
-};
